--- .dockerignore
@@ -1,42 +1,0 @@
-# Dependencies
-node_modules
-npm-debug.log*
-yarn-debug.log*
-yarn-error.log*
-
-# Testing
-coverage
-.nyc_output
-__tests__
-
-# Next.js
-.next
-out
-dist
-build
-
-# Production
-.env*.local
-.env.production
-
-# Misc
-.DS_Store
-*.pem
-.vscode
-.idea
-.kiro
-
-# Debug
-*.log
-logs
-
-# Git
-.git
-.gitignore
-
-# Documentation
-*.md
-!README.md
-
-# Test files
-Tests&fixes
