--- .vscode/settings.json
@@ -1,4 +1,0 @@
-{
-    "IDX.aI.enableInlineCompletion": true,
-    "IDX.aI.enableCodebaseIndexing": true
-}
