--- Tests&fixes/ISSUES_FOUND.md
@@ -1,143 +1,0 @@
-# AIRLAB Website - Issues & Fixes Report
-
-## 🎯 Executive Summary
-
-**Total Issues Found:** 15  
-**Critical Issues:** 3 (1 fixed, 2 documented)  
-**Fixed Issues:** 7 ✅  
-**Documented Issues:** 8 ⚠️  
-**Test Files Created:** 3  
-**Documentation Files:** 9
-
----
-
-## Critical Issues Found
-
-### 1. Missing Dependency: framer-motion
-**Location:** `src/components/layout/PageWrapper.tsx`
-**Issue:** Component imports `framer-motion` but it's not in package.json
-**Impact:** Build will fail
-**Status:** ✅ FIXED - Removed unused import
-
-### 2. Security Issues in API Routes
-**Location:** All `/src/app/api/admin/*` routes
-**Issues:**
-- No authentication/authorization checks
-- Direct file system writes without validation
-- No input sanitization
-- Race conditions in file operations
-**Impact:** Critical security vulnerability - anyone can modify data
-**Status:** ⚠️ NEEDS ATTENTION
-
-### 3. Inconsistent ID Generation
-**Location:** `src/app/api/admin/news/route.ts` vs `src/app/api/admin/projects/route.ts`
-**Issue:** News uses numeric IDs, projects use string IDs
-**Impact:** Data inconsistency
-**Status:** ✅ FIXED
-
-### 4. Social Media Link Error
-**Location:** `src/data/socials.ts`
-**Issue:** Twitter/X link is incomplete: "x.com" instead of "https://x.com/..."
-**Impact:** Broken link
-**Status:** ✅ FIXED
-
-### 5. Console.log in Production Code
-**Location:** `src/app/contact/page.tsx`
-**Issue:** console.log statement in form submission
-**Impact:** Minor - will be removed in production build
-**Status:** ✅ FIXED
-
-### 6. Redundant Admin Routes
-**Location:** `/admin` and `/admin-air-airlabalaba`
-**Issue:** Two admin routes doing the same thing, confusing naming
-**Impact:** Maintenance overhead, confusing UX
-**Status:** ⚠️ NEEDS DECISION
-
-### 7. Missing Animation Classes
-**Location:** `src/components/layout/PageWrapper.tsx`
-**Issue:** Uses non-existent utility classes: `animation-delay-200`, `animation-fill-forwards`
-**Impact:** Animations won't work as expected
-**Status:** ✅ FIXED
-
-### 8. TypeScript Configuration Issues
-**Location:** `next.config.ts`
-**Issue:** `ignoreBuildErrors: true` and `ignoreDuringBuilds: true`
-**Impact:** Hides real errors, technical debt
-**Status:** ⚠️ SHOULD BE REMOVED AFTER FIXING TYPES
-
-### 9. Missing News IDs
-**Location:** `src/data/news.json`
-**Issue:** News items don't have IDs but API expects them
-**Impact:** Delete/Update operations will fail
-**Status:** ✅ FIXED
-
-### 10. Incomplete Project Data
-**Location:** `src/data/projects.json`
-**Issue:** Only 3 projects but dashboard shows "4"
-**Impact:** Incorrect statistics
-**Status:** ⚠️ NEEDS DATA UPDATE
-
-## Code Quality Issues
-
-### 11. Unused Import
-**Location:** `src/components/layout/PageWrapper.tsx`
-**Issue:** Imports `motion` from framer-motion but never uses it
-**Status:** ✅ FIXED
-
-### 12. Hardcoded Statistics
-**Location:** `src/app/admin-air-airlabalaba/page.tsx`
-**Issue:** Dashboard stats are hardcoded, not dynamic
-**Impact:** Stats won't update automatically
-**Status:** ⚠️ SHOULD BE DYNAMIC
-
-### 13. Missing Error Handling
-**Location:** Multiple API routes
-**Issue:** Generic error messages, no proper error logging
-**Status:** ⚠️ NEEDS IMPROVEMENT
-
-## Performance Issues
-
-### 14. No Image Optimization
-**Location:** Various pages
-**Issue:** Using external images without optimization
-**Status:** ✅ Already configured in next.config.ts
-
-### 15. No Loading States
-**Location:** Admin pages
-**Issue:** No loading indicators for async operations
-**Status:** ⚠️ NEEDS IMPROVEMENT
-
-## Testing Recommendations
-
-### Unit Tests Needed:
-- [ ] API route handlers
-- [ ] Form validation
-- [ ] Data transformation utilities
-- [ ] Component rendering
-
-### Integration Tests Needed:
-- [ ] Admin CRUD operations
-- [ ] Navigation flow
-- [ ] Form submissions
-- [ ] Theme switching
-
-### E2E Tests Needed:
-- [ ] Complete user journey
-- [ ] Admin workflow
-- [ ] Mobile responsiveness
-- [ ] Cross-browser compatibility
-
-## Fixed Issues Summary
-✅ Removed framer-motion import
-✅ Fixed social media links
-✅ Removed console.log
-✅ Fixed animation classes
-✅ Added IDs to news items
-✅ Standardized ID generation
-
-## Remaining Issues
-⚠️ API authentication/authorization
-⚠️ Redundant admin routes
-⚠️ TypeScript error suppression
-⚠️ Hardcoded dashboard stats
-⚠️ Missing loading states
