--- docker-compose.yml
@@ -1,32 +1,0 @@
-version: '3.8'
-
-services:
-  airlab-website:
-    build:
-      context: .
-      dockerfile: Dockerfile
-    container_name: airlab-website
-    restart: unless-stopped
-    ports:
-      - "3000:3000"
-    environment:
-      - NODE_ENV=production
-      - ADMIN_PASSWORD=${ADMIN_PASSWORD}
-      - JWT_SECRET=${JWT_SECRET}
-      - NEXT_PUBLIC_SITE_URL=${NEXT_PUBLIC_SITE_URL}
-      - NEXT_PUBLIC_SITE_NAME=${NEXT_PUBLIC_SITE_NAME}
-      - NEXT_PUBLIC_CONTACT_EMAIL=${NEXT_PUBLIC_CONTACT_EMAIL}
-    volumes:
-      - ./src/data:/app/src/data
-    networks:
-      - airlab-network
-    healthcheck:
-      test: ["CMD", "wget", "--quiet", "--tries=1", "--spider", "http://localhost:3000"]
-      interval: 30s
-      timeout: 10s
-      retries: 3
-      start_period: 40s
-
-networks:
-  airlab-network:
-    driver: bridge
