--- docs/blueprint.md
@@ -1,19 +1,0 @@
-# **App Name**: AIROL Unilag
-
-## Core Features:
-
-- Interactive 3D Elements: Interactive 3D elements on the homepage to showcase lab projects and robotics. Use floating elements to give a modern and dynamic feel.
-- Dynamic Page Layouts: Unique page designs for each section (About, Research, Projects, Team, Contact) with creative layouts and sliding animations for navigation. Each section to clearly show content regarding the AI and Robotics Lab at UNILAG.
-- Theme Toggle: Light and dark mode switch to improve accessibility and user preference. Retains design and readability in each mode.
-- Admin CMS: Admin dashboard for uploading and managing content (projects, research papers, team member profiles, news, events). The CMS will be tailored so the content remains within design boundaries and guidelines.
-- Paper Summarizer: AI-powered tool to summarize the content for each of the uploaded research papers.
-
-## Style Guidelines:
-
-- Primary color: Soft Green (#A7D1AB) to subtly evoke the lab's green and white logo. Not overtly 'artificial intelligence'-themed; promotes freshness and approachability.
-- Background color: Dark Grey (#2B302D) for a modern, sophisticated feel and contrast with interactive elements. Promotes content readability.
-- Accent color: Teal (#6FB5B8) for highlights, interactive elements, and call-to-action buttons.
-- Headline font: 'Belleza', a humanist sans-serif for headers and titles to convey art and design. Body font: 'Alegreya', a humanist serif to facilitate the reading of long sections of text.
-- Modern, minimalist icons related to AI, robotics, and research, in green and white outlines.
-- Asymmetrical layouts and generous spacing to give a modern, open, and engaging feel.
-- Smooth transitions and subtle animations on scroll and hover, to create an engaging experience without overwhelming the user.
