--- docs/changelog.md
@@ -1,204 +1,0 @@
-# AIRLAB Website - Recent Updates
-
-## Branding Updates ✅
-
-### Name Standardization
-- **Changed from**: "AIRLAB" / "AIROL" (inconsistent)
-- **Changed to**: "AIRLAB" (Artificial Intelligence & Robotics Laboratory)
-- Updated across all pages, titles, and content
-
-### Page Title
-- **New**: "AIRLAB - Artificial Intelligence & Robotics Laboratory | University of Lagos"
-- Improved SEO with descriptive, keyword-rich title
-
-## Content Improvements ✅
-
-### Homepage
-- **Hero Section**: Updated copy to be more impactful
-  - New headline: "Pioneering the Future of Intelligent Systems"
-  - Enhanced description emphasizing Africa-focused solutions
-- **Lab Coordinator Section**: Improved bio for Dr. Chika Yinka-Banjo
-- **Focus Areas**: Better descriptions of research, projects, and collaboration
-- **Get Involved Section**: More specific and actionable descriptions
-
-### Research Page
-- **Added**: 6 dummy research paper cards with realistic data
-- **Improved**: Layout now uses 3-column grid
-- **Enhanced**: Better visual hierarchy with year badges
-- **Fixed**: Now displays actual content instead of empty state
-
-### About Page
-- Updated all references to "AIRLAB"
-- Corrected "FLL" capitalization
-- Better mission and vision statements
-
-### Other Pages
-- Team, Projects, and other pages updated with consistent branding
-
-## Data Structure Changes ✅
-
-### New JSON Data Files
-Created centralized data management in `src/data/`:
-
-1. **`projects.json`** (4 projects)
-   - AI-Powered Traffic Management System
-   - Robotic Assistant for Elderly Care
-   - Precision Agriculture Drone
-   - Natural Language Interface for Databases
-
-2. **`team.json`** (13 team members)
-   - Dr. Chika Yinka-Banjo (Director)
-   - Mrs Omokhaba Blessing Yama (Senior Researcher)
-   - 11 Masters/PhD students with unique bios
-
-3. **`research.json`** (6 research papers)
-   - Deep Learning for Medical Image Analysis
-   - Autonomous Navigation in Unstructured Environments
-   - NLP for Low-Resource Languages
-   - Ethical AI: Bias Detection
-   - Smart Grid Optimization
-   - Human-Robot Interaction in Education
-
-### Benefits
-- Easy content updates without touching code
-- Consistent data structure across the site
-- Single source of truth for all content
-- Simple for non-technical team members to update
-
-## Admin Panel Improvements ✅
-
-### Password Protection
-- **Password**: `AIRLAB_2025`
-- Session-based authentication
-- Professional login UI with lock icon
-
-### Research Management Page
-- **Added**: Search functionality (works in real-time)
-- **Added**: Data loading from JSON file
-- **Added**: Delete functionality with toast notifications
-- **Fixed**: Now displays actual research papers
-- **Improved**: Better empty state handling
-
-### Upload Page
-- **Improved**: Better form styling with theme consistency
-- **Added**: File name display after selection
-- **Added**: Cancel button
-- **Added**: Form validation
-- **Added**: Success feedback and auto-redirect
-- **Enhanced**: Better UX with clear labels and placeholders
-
-### Dashboard
-- **Updated**: Statistics to match actual data counts
-- **Updated**: Recent activity with realistic examples
-- **Maintained**: Quick action buttons and overview cards
-
-## Image Updates ✅
-
-### Current Status
-- Using high-quality Unsplash images as placeholders
-- All images are properly sized and optimized
-- Better alt text for accessibility
-
-### Recommendations
-- See `docs/image-recommendations.md` for detailed guide
-- Priority: Replace with actual AIRLAB photos
-- 13 team member photos needed
-- Lab interior shots recommended
-- Project and research photos would enhance authenticity
-
-## Technical Improvements ✅
-
-### Code Quality
-- Removed all backend dependencies
-- Clean, maintainable component structure
-- Proper TypeScript typing
-- No diagnostic errors
-
-### Performance
-- Optimized images with proper sizing
-- Efficient data loading
-- Fast page transitions
-
-### Accessibility
-- Proper alt text on images
-- Semantic HTML structure
-- Keyboard navigation support
-- ARIA labels where needed
-
-## What Still Needs Work
-
-### High Priority
-1. **Replace placeholder images** with actual AIRLAB photos
-2. **Update team member bios** with accurate information
-3. **Add real project details** and links
-4. **Update research paper links** to actual PDFs
-
-### Medium Priority
-1. **Contact page** - Add functional contact form or email
-2. **News section** - Add actual news items
-3. **Project detail pages** - Create individual project pages
-4. **Team member detail pages** - Expand on individual profiles
-
-### Low Priority
-1. **Analytics** - Add Google Analytics or similar
-2. **SEO** - Add meta descriptions for all pages
-3. **Social sharing** - Add Open Graph tags
-4. **Blog section** - Consider adding a blog for updates
-
-## How to Update Content
-
-### To Update Team Members:
-1. Open `src/data/team.json`
-2. Edit the relevant team member object
-3. Save the file - changes appear immediately
-
-### To Update Projects:
-1. Open `src/data/projects.json`
-2. Add/edit project objects
-3. Save the file
-
-### To Update Research Papers:
-1. Open `src/data/research.json`
-2. Add/edit research paper objects
-3. Upload PDF files to `/public/papers/`
-4. Update `fileUrl` to point to the PDF
-
-### To Update Images:
-1. Place images in `/public/images/`
-2. Update the `imageUrl` in the relevant JSON file
-3. Use relative paths: `/images/team/john-doe.jpg`
-
-## Admin Panel Usage
-
-### Accessing Admin:
-1. Navigate to `/admin`
-2. Enter password: `AIRLAB_2025`
-3. Access granted for the session
-
-### Managing Research Papers:
-1. Go to Admin > Research Papers
-2. Use search to find specific papers
-3. Click trash icon to remove papers
-4. Click "Upload New Paper" to add papers
-
-### Note on Admin Functionality:
-- Admin panel is currently **frontend-only**
-- Changes are **not persisted** (refresh resets data)
-- To make changes permanent, edit the JSON files directly
-- For full CMS functionality, a backend would be needed
-
-## Summary
-
-The website now has:
-- ✅ Consistent "AIRLAB" branding throughout
-- ✅ Professional, engaging copy
-- ✅ Structured JSON data for easy updates
-- ✅ Functional admin panel with password protection
-- ✅ 6 research papers displayed
-- ✅ 13 team members with unique profiles
-- ✅ 4 detailed projects
-- ✅ Better images (placeholders ready for replacement)
-- ✅ Improved UX across all pages
-- ✅ No errors or broken functionality
-
-Next steps: Replace placeholder images with actual AIRLAB photos and update content with real information.
