--- ecosystem.config.js
@@ -1,42 +1,0 @@
-// PM2 Ecosystem Configuration for AIRLAB Website
-module.exports = {
-  apps: [
-    {
-      name: 'airlab-website',
-      script: 'npm',
-      args: 'start',
-      cwd: './',
-      instances: 1,
-      autorestart: true,
-      watch: false,
-      max_memory_restart: '1G',
-      env: {
-        NODE_ENV: 'production',
-        PORT: 3000,
-      },
-      env_production: {
-        NODE_ENV: 'production',
-        PORT: 3000,
-      },
-      error_file: './logs/pm2-error.log',
-      out_file: './logs/pm2-out.log',
-      log_file: './logs/pm2-combined.log',
-      time: true,
-      merge_logs: true,
-      log_date_format: 'YYYY-MM-DD HH:mm:ss Z',
-    },
-  ],
-
-  deploy: {
-    production: {
-      user: 'www-data',
-      host: 'your-server-ip',
-      ref: 'origin/main',
-      repo: 'git@github.com:your-repo/AIRlab-site.git',
-      path: '/var/www/airlab-website',
-      'pre-deploy-local': '',
-      'post-deploy': 'npm install --legacy-peer-deps && npm run build && pm2 reload ecosystem.config.js --env production',
-      'pre-setup': '',
-    },
-  },
-};
