--- src/app/admin-air-airlabalaba/team/page.tsx
@@ -1,373 +1,0 @@
-"use client";
-
-import { Card, CardContent, CardDescription, CardHeader, CardTitle } from '@/components/ui/card';
-import { Button } from '@/components/ui/button';
-import Link from 'next/link';
-import { PlusCircle, Edit, Trash2, Search, Users } from 'lucide-react';
-import { Input } from '@/components/ui/input';
-import { Table, TableBody, TableCell, TableHead, TableHeader, TableRow } from '@/components/ui/table';
-import { Avatar, AvatarFallback, AvatarImage } from '@/components/ui/avatar';
-import { useState, useEffect } from 'react';
-import { useToast } from '@/hooks/use-toast';
-import { Skeleton } from '@/components/ui/skeleton';
-import {
-  Dialog,
-  DialogContent,
-  DialogDescription,
-  DialogFooter,
-  DialogHeader,
-  DialogTitle,
-} from '@/components/ui/dialog';
-import { Label } from '@/components/ui/label';
-import { Textarea } from '@/components/ui/textarea';
-import {
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
-} from '@/components/ui/select';
-
-interface TeamMember {
-  id: string;
-  name: string;
-  role: string;
-  imageUrl: string;
-  bio?: string;
-  social?: {
-    linkedin?: string;
-    twitter?: string;
-    github?: string;
-    email?: string;
-  };
-}
-
-function TeamTableSkeleton() {
-  return (
-    <div className="space-y-4">
-      {Array.from({ length: 5 }).map((_, i) => (
-        <div key={i} className="flex items-center space-x-4 p-4">
-          <Skeleton className="h-10 w-10 rounded-full" />
-          <Skeleton className="h-4 w-32" />
-          <Skeleton className="h-4 w-24" />
-          <Skeleton className="h-4 w-40" />
-          <div className="flex space-x-2 ml-auto">
-            <Skeleton className="h-8 w-8" />
-            <Skeleton className="h-8 w-8" />
-          </div>
-        </div>
-      ))}
-    </div>
-  );
-}
-
-export default function AdminTeamPage() {
-  const [teamMembers, setTeamMembers] = useState<TeamMember[]>([]);
-  const [loading, setLoading] = useState(true);
-  const [searchTerm, setSearchTerm] = useState('');
-  const [selectedCategory, setSelectedCategory] = useState<string>('leading');
-  const [editingMember, setEditingMember] = useState<TeamMember | null>(null);
-  const [isDialogOpen, setIsDialogOpen] = useState(false);
-  const { toast } = useToast();
-
-  useEffect(() => {
-    fetchTeamMembers();
-  }, [selectedCategory]);
-
-  const fetchTeamMembers = async () => {
-    setLoading(true);
-    try {
-      const response = await fetch(`/api/admin/team?category=${selectedCategory}`);
-      if (response.ok) {
-        const result = await response.json();
-        setTeamMembers(result.data || []);
-      } else {
-        toast({
-          title: "Error",
-          description: "Failed to fetch team members",
-          variant: "destructive",
-        });
-      }
-    } catch {
-      toast({
-        title: "Error",
-        description: "Failed to fetch team members",
-        variant: "destructive",
-      });
-    } finally {
-      setLoading(false);
-    }
-  };
-
-  const handleEdit = (member: TeamMember) => {
-    setEditingMember(member);
-    setIsDialogOpen(true);
-  };
-
-  const handleSave = async () => {
-    if (!editingMember) return;
-
-    try {
-      const response = await fetch('/api/admin/team', {
-        method: 'PUT',
-        headers: {
-          'Content-Type': 'application/json',
-        },
-        body: JSON.stringify({ ...editingMember, category: selectedCategory }),
-      });
-
-      if (response.ok) {
-        await fetchTeamMembers();
-        setIsDialogOpen(false);
-        setEditingMember(null);
-        toast({
-          title: "Success",
-          description: "Team member updated successfully",
-        });
-      } else {
-        toast({
-          title: "Error",
-          description: "Failed to update team member",
-          variant: "destructive",
-        });
-      }
-    } catch {
-      toast({
-        title: "Error",
-        description: "Failed to update team member",
-        variant: "destructive",
-      });
-    }
-  };
-
-  const handleDelete = async (id: string) => {
-    if (!confirm('Are you sure you want to delete this team member?')) return;
-
-    try {
-      const response = await fetch(`/api/admin/team?id=${id}&category=${selectedCategory}`, {
-        method: 'DELETE',
-      });
-
-      if (response.ok) {
-        await fetchTeamMembers();
-        toast({
-          title: "Success",
-          description: "Team member deleted successfully",
-        });
-      } else {
-        toast({
-          title: "Error",
-          description: "Failed to delete team member",
-          variant: "destructive",
-        });
-      }
-    } catch {
-      toast({
-        title: "Error",
-        description: "Failed to delete team member",
-        variant: "destructive",
-      });
-    }
-  };
-
-  const filteredMembers = teamMembers.filter(member =>
-    member.name.toLowerCase().includes(searchTerm.toLowerCase()) ||
-    member.role.toLowerCase().includes(searchTerm.toLowerCase()) ||
-    (member.social?.email || '').toLowerCase().includes(searchTerm.toLowerCase())
-  );
-
-  return (
-    <div className="space-y-6">
-      <div className="flex items-center justify-between">
-        <div className="flex items-center gap-4">
-          <div>
-            <h2 className="font-headline text-3xl font-semibold flex items-center">
-              <Users className="mr-3 h-8 w-8 text-primary" /> Manage Team Members
-            </h2>
-            <p className="text-muted-foreground font-body">Add, edit, or remove team member profiles.</p>
-          </div>
-        </div>
-        <Button asChild className="bg-primary hover:bg-primary/90 text-primary-foreground">
-          <Link href="/admin-air-airlabalaba/team/new">
-            <PlusCircle className="mr-2 h-5 w-5" /> Add Team Member
-          </Link>
-        </Button>
-      </div>
-
-      <Card className="shadow-lg">
-        <CardHeader>
-          <CardTitle className="font-headline">Team Member List</CardTitle>
-          <CardDescription className="font-body">
-            All team members in the system.
-            <div className="mt-4 flex items-center gap-4">
-              <Search className="h-5 w-5 text-muted-foreground" />
-              <Input 
-                placeholder="Search members..." 
-                className="max-w-sm"
-                value={searchTerm}
-                onChange={(e) => setSearchTerm(e.target.value)}
-              />
-              <Select value={selectedCategory} onValueChange={setSelectedCategory}>
-                <SelectTrigger className="w-48">
-                  <SelectValue placeholder="Select category" />
-                </SelectTrigger>
-                <SelectContent>
-                  <SelectItem value="leading">Leading Team</SelectItem>
-                  <SelectItem value="pioneer">Pioneer Team</SelectItem>
-                  <SelectItem value="volunteers">Volunteers</SelectItem>
-                </SelectContent>
-              </Select>
-            </div>
-          </CardDescription>
-        </CardHeader>
-        <CardContent>
-          {loading ? (
-            <TeamTableSkeleton />
-          ) : (
-            <Table>
-              <TableHeader>
-                <TableRow>
-                  <TableHead>Avatar</TableHead>
-                  <TableHead>Name</TableHead>
-                  <TableHead>Role</TableHead>
-                  <TableHead>Email</TableHead>
-                  <TableHead className="text-right">Actions</TableHead>
-                </TableRow>
-              </TableHeader>
-              <TableBody>
-                {filteredMembers.length === 0 ? (
-                  <TableRow>
-                    <TableCell colSpan={5} className="text-center text-muted-foreground py-8">
-                      {searchTerm ? 'No members found matching your search.' : 'No team members available.'}
-                    </TableCell>
-                  </TableRow>
-                ) : (
-                  filteredMembers.map((member) => (
-                  <TableRow key={member.id}>
-                    <TableCell>
-                      <Avatar>
-                        <AvatarImage src={member.imageUrl} alt={member.name} />
-                        <AvatarFallback>{member.name.substring(0,2).toUpperCase()}</AvatarFallback>
-                      </Avatar>
-                    </TableCell>
-                    <TableCell className="font-medium">{member.name}</TableCell>
-                    <TableCell>{member.role}</TableCell>
-                    <TableCell>{member.social?.email || '-'}</TableCell>
-                    <TableCell className="text-right">
-                      <div className="flex justify-end gap-2">
-                        <Button 
-                          variant="outline" 
-                          size="sm"
-                          onClick={() => handleEdit(member)}
-                        >
-                          <Edit className="h-4 w-4 mr-1" />
-                          Edit
-                        </Button>
-                        <Button 
-                          variant="destructive" 
-                          size="sm"
-                          onClick={() => handleDelete(member.id)}
-                        >
-                          <Trash2 className="h-4 w-4 mr-1" />
-                          Delete
-                        </Button>
-                      </div>
-                    </TableCell>
-                  </TableRow>
-                )))}
-              </TableBody>
-            </Table>
-          )}
-        </CardContent>
-      </Card>
-
-      <Dialog open={isDialogOpen} onOpenChange={setIsDialogOpen}>
-        <DialogContent className="max-w-2xl">
-          <DialogHeader>
-            <DialogTitle>Edit Team Member</DialogTitle>
-            <DialogDescription>
-              Make changes to the team member details below.
-            </DialogDescription>
-          </DialogHeader>
-          {editingMember && (
-            <div className="grid gap-4 py-4">
-              <div className="grid grid-cols-4 items-center gap-4">
-                <Label htmlFor="name" className="text-right">
-                  Name
-                </Label>
-                <Input
-                  id="name"
-                  value={editingMember.name}
-                  onChange={(e) => setEditingMember({...editingMember, name: e.target.value})}
-                  className="col-span-3"
-                />
-              </div>
-              <div className="grid grid-cols-4 items-center gap-4">
-                <Label htmlFor="role" className="text-right">
-                  Role
-                </Label>
-                <Input
-                  id="role"
-                  value={editingMember.role}
-                  onChange={(e) => setEditingMember({...editingMember, role: e.target.value})}
-                  className="col-span-3"
-                />
-              </div>
-              <div className="grid grid-cols-4 items-center gap-4">
-                <Label htmlFor="imageUrl" className="text-right">
-                  Image URL
-                </Label>
-                <Input
-                  id="imageUrl"
-                  value={editingMember.imageUrl}
-                  onChange={(e) => setEditingMember({...editingMember, imageUrl: e.target.value})}
-                  className="col-span-3"
-                />
-              </div>
-              <div className="grid grid-cols-4 items-center gap-4">
-                <Label htmlFor="bio" className="text-right">
-                  Bio
-                </Label>
-                <Textarea
-                  id="bio"
-                  value={editingMember.bio || ''}
-                  onChange={(e) => setEditingMember({...editingMember, bio: e.target.value})}
-                  className="col-span-3"
-                />
-              </div>
-              <div className="grid grid-cols-4 items-center gap-4">
-                <Label htmlFor="email" className="text-right">
-                  Email
-                </Label>
-                <Input
-                  id="email"
-                  type="email"
-                  value={editingMember.social?.email || ''}
-                  onChange={(e) => setEditingMember({...editingMember, social: {...editingMember.social, email: e.target.value}})}
-                  className="col-span-3"
-                />
-              </div>
-              <div className="grid grid-cols-4 items-center gap-4">
-                <Label htmlFor="linkedin" className="text-right">
-                  LinkedIn
-                </Label>
-                <Input
-                  id="linkedin"
-                  value={editingMember.social?.linkedin || ''}
-                  onChange={(e) => setEditingMember({...editingMember, social: {...editingMember.social, linkedin: e.target.value}})}
-                  className="col-span-3"
-                  placeholder="https://linkedin.com/in/..."
-                />
-              </div>
-            </div>
-          )}
-          <DialogFooter>
-            <Button type="submit" onClick={handleSave}>
-              Save changes
-            </Button>
-          </DialogFooter>
-        </DialogContent>
-      </Dialog>
-    </div>
-  );
-}
